Move `ArrayVariantToArrowRowBuilder` and `VariantToListArrowRowBuilder` to `variant_to_arrow`
